requirements.txt
Flask
flask-cors
openstacksdk
python-dotenv
schedule